refactor: move BIOhazard
